multilingual_chatbot_python/requirements.txt
openai
python-dotenv
langdetect
streamlit
audio_recorder_streamlit
streamlit-float
